Issue #8: Rename tester to Fire Arrow Server UI and restore functionality
- Rename FhirTesterConfigCondition to FireArrowServerConfigCondition
- Rename FhirTesterConfig to FireArrowServerConfig
- Add EnvironmentHelper.isFireArrowServerUiEnabled() method
- Update condition to check for fire_arrow_server configuration
- Restore UI functionality with Fire Arrow Server branding
- Remove temporary home controller (using proper tester UI instead)
- UI now shows when fire_arrow_server config is present in application.yaml

This restores the proven tester UI functionality but with Fire Arrow Server
branding and configuration, fixing the 404 error on root path.
